requirements.txt
appdirs==1.4.4
atomos==0.3.1
boto3==1.11.7
botocore==1.14.7
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
chevron==0.13.1
distlib==0.3.1
dnspython==1.16.0
docutils==0.15.2
filelock==3.0.12
google-auth==1.10.0
idna==2.8
jmespath==0.9.4
kubernetes==10.0.1
mock==3.0.5
munch==2.5.0
oauthlib==3.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
tox==3.23.1
urllib3==1.25.7
virtualenv==20.4.7
websocket-client==0.56.0
